pyproject.toml
[project]
name = "to-markdown"
version = "0.1.0"
description = "CLI file-to-Markdown converter optimized for LLM consumption"
readme = "README.md"
requires-python = ">=3.14"
license = "MIT"
dependencies = [
  "kreuzberg>=4.3.8",
  "typer>=0.24.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
llm = [
  "google-genai>=1.59.0",
  "tenacity>=8.0",
  "python-dotenv>=1.0",
]
dev = [
  "ruff>=0.15.2",
  "pytest>=9.0",
  "pytest-cov>=6.0",
  "syrupy>=5.0",
  "fpdf2>=2.8",
  "python-docx>=1.1",
  "openpyxl>=3.1",
  "python-pptx>=1.0",
  "Pillow>=11.0",
]

[project.scripts]
to-markdown = "to_markdown.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/to_markdown"]

[tool.ruff]
target-version = "py314"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "ocr: tests requiring Tesseract OCR (skipped if not installed)",
]
